Creates discrete namespaces for each of PSR-6 and PSR-16 implementations
Creates `Zend\Cache\Psr\CacheItemPool` and `Zend\Cache\Psr\SimpleCache`
namespaces, moving class files for each implementation under the
appropriate namespace.
