+++ mkdocs.yml
@@ -13,8 +13,9 @@ pages:
         - ObjectCache: pattern/object-cache.md
         - OutputCache: pattern/output-cache.md
         - CaptureCache: pattern/capture-cache.md
-    - PSR-6: psr6.md
-    - PSR-16: psr16.md
+    - "Standards Support":
+        - PSR-6: psr6.md
+        - PSR-16: psr16.md
 site_name: zend-cache
 site_description: Zend\Cache
 repo_url: 'https://github.com/zendframework/zend-cache'
