--- src/Psr/CacheException.php
@@ -1,16 +1,0 @@
-<?php
-/**
- * Zend Framework (http://framework.zend.com/).
- *
- * @link      http://github.com/zendframework/zend-cache for the canonical source repository
- * @copyright Copyright (c) 2016 Zend Technologies USA Inc. (http://www.zend.com)
- * @license   http://framework.zend.com/license/new-bsd New BSD License
- */
-
-namespace Zend\Cache\Psr;
-
-use Psr\Cache\CacheException as CacheExceptionInterface;
-
-class CacheException extends \RuntimeException implements CacheExceptionInterface
-{
-}
