+++ src/Psr/SimpleCache/SimpleCacheInvalidArgumentException.php
@@ -5,10 +5,11 @@
  * @license   https://github.com/zendframework/zend-cache/blob/master/LICENSE.md New BSD License
  */
 
-namespace Zend\Cache\Psr;
+namespace Zend\Cache\Psr\SimpleCache;
 
+use InvalidArgumentException;
 use Psr\SimpleCache\InvalidArgumentException as PsrInvalidArgumentException;
 
-class SimpleCacheInvalidArgumentException extends \InvalidArgumentException implements PsrInvalidArgumentException
+class SimpleCacheInvalidArgumentException extends InvalidArgumentException implements PsrInvalidArgumentException
 {
 }
