+++ test/Psr/CacheItemPool/ApcIntegrationTest.php
@@ -1,22 +1,22 @@
 <?php
 /**
- * Zend Framework (http://framework.zend.com/).
- *
- * @link      http://github.com/zendframework/zend-cache for the canonical source repository
- * @copyright Copyright (c) 2016 Zend Technologies USA Inc. (http://www.zend.com)
- * @license   http://framework.zend.com/license/new-bsd New BSD License
+ * @see       https://github.com/zendframework/zend-cache for the canonical source repository
+ * @copyright Copyright (c) 2018 Zend Technologies USA Inc. (http://www.zend.com)
+ * @license   https://github.com/zendframework/zend-cache/blob/master/LICENSE.md New BSD License
  */
 
-namespace ZendTest\Cache\Psr;
+namespace ZendTest\Cache\Psr\CacheItemPool;
 
 use Cache\IntegrationTests\CachePoolTest;
-use Zend\Cache\Psr\CacheItemPoolAdapter;
-use Zend\Cache\Storage\Adapter\MongoDb;
+use Zend\Cache\Psr\CacheItemPool\CacheItemPoolAdapter;
 use Zend\Cache\StorageFactory;
 use Zend\Cache\Exception;
 use Zend\ServiceManager\Exception\ServiceNotCreatedException;
 
-class MongoDbIntegrationTest extends CachePoolTest
+/**
+ * @requires extension apcu
+ */
+class ApcIntegrationTest extends CachePoolTest
 {
     /**
      * Backup default timezone
@@ -25,19 +25,25 @@ class MongoDbIntegrationTest extends CachePoolTest
     private $tz;
 
     /**
-     * @var MongoDb
+     * Restore 'apc.use_request_time'
+     *
+     * @var mixed
      */
-    private $storage;
+    protected $iniUseRequestTime;
 
     protected function setUp()
     {
-        if (! getenv('TESTS_ZEND_CACHE_MONGODB_ENABLED')) {
-            $this->markTestSkipped('Enable TESTS_ZEND_CACHE_MONGODB_ENABLED to run this test');
+        if (! getenv('TESTS_ZEND_CACHE_APC_ENABLED')) {
+            $this->markTestSkipped('Enable TESTS_ZEND_CACHE_APC_ENABLED to run this test');
         }
 
         // set non-UTC timezone
         $this->tz = date_default_timezone_get();
         date_default_timezone_set('America/Vancouver');
+
+        // needed on test expirations
+        $this->iniUseRequestTime = ini_get('apc.use_request_time');
+        ini_set('apc.use_request_time', 0);
 
         parent::setUp();
     }
@@ -46,21 +52,29 @@ class MongoDbIntegrationTest extends CachePoolTest
     {
         date_default_timezone_set($this->tz);
 
-        if ($this->storage) {
-            $this->storage->flush();
+        if (function_exists('apc_clear_cache')) {
+            apc_clear_cache('user');
         }
 
+        // reset ini configurations
+        ini_set('apc.use_request_time', $this->iniUseRequestTime);
+
         parent::tearDown();
+    }
+
+    /**
+     * @expectedException \Zend\Cache\Psr\CacheItemPool\CacheException
+     */
+    public function testApcUseRequestTimeThrowsException()
+    {
+        ini_set('apc.use_request_time', 1);
+        $this->createCachePool();
     }
 
     public function createCachePool()
     {
         try {
-            $storage = StorageFactory::adapterFactory('mongodb', [
-                'server'     => getenv('TESTS_ZEND_CACHE_MONGODB_CONNECTSTRING'),
-                'database'   => getenv('TESTS_ZEND_CACHE_MONGODB_DATABASE'),
-                'collection' => getenv('TESTS_ZEND_CACHE_MONGODB_COLLECTION'),
-            ]);
+            $storage = StorageFactory::adapterFactory('apc');
 
             $deferredSkippedMessage = sprintf(
                 '%s storage doesn\'t support driver deferred',
