+++ test/Psr/CacheItemPool/BlackHoleIntegrationTest.php
@@ -1,0 +1,24 @@
+<?php
+/**
+ * @see       https://github.com/zendframework/zend-cache for the canonical source repository
+ * @copyright Copyright (c) 2018 Zend Technologies USA Inc. (http://www.zend.com)
+ * @license   https://github.com/zendframework/zend-cache/blob/master/LICENSE.md New BSD License
+ */
+
+namespace ZendTest\Cache\Psr\CacheItemPool;
+
+use PHPUnit\Framework\TestCase;
+use Zend\Cache\Psr\CacheItemPool\CacheItemPoolAdapter;
+use Zend\Cache\StorageFactory;
+
+class BlackHoleIntegrationTest extends TestCase
+{
+    /**
+     * @expectedException \Zend\Cache\Psr\CacheItemPool\CacheException
+     */
+    public function testAdapterNotSupported()
+    {
+        $storage = StorageFactory::adapterFactory('blackhole');
+        new CacheItemPoolAdapter($storage);
+    }
+}
