+++ test/Psr/CacheItemPool/CacheItemTest.php
@@ -1,18 +1,16 @@
 <?php
 /**
- * Zend Framework (http://framework.zend.com/).
- *
- * @link      http://github.com/zendframework/zend-cache for the canonical source repository
- * @copyright Copyright (c) 2016 Zend Technologies USA Inc. (http://www.zend.com)
- * @license   http://framework.zend.com/license/new-bsd New BSD License
+ * @see       https://github.com/zendframework/zend-cache for the canonical source repository
+ * @copyright Copyright (c) 2018 Zend Technologies USA Inc. (http://www.zend.com)
+ * @license   https://github.com/zendframework/zend-cache/blob/master/LICENSE.md New BSD License
  */
 
-namespace ZendTest\Cache\Psr;
+namespace ZendTest\Cache\Psr\CacheItemPool;
 
 use DateInterval;
 use DateTime;
 use PHPUnit\Framework\TestCase;
-use Zend\Cache\Psr\CacheItem;
+use Zend\Cache\Psr\CacheItemPool\CacheItem;
 
 class CacheItemTest extends TestCase
 {
@@ -72,7 +70,7 @@ class CacheItemTest extends TestCase
     }
 
     /**
-     * @expectedException \Zend\Cache\Psr\InvalidArgumentException
+     * @expectedException \Zend\Cache\Psr\CacheItemPool\InvalidArgumentException
      */
     public function testExpireAtInvalidThrowsException()
     {
@@ -105,7 +103,7 @@ class CacheItemTest extends TestCase
     }
 
     /**
-     * @expectedException \Zend\Cache\Psr\InvalidArgumentException
+     * @expectedException \Zend\Cache\Psr\CacheItemPool\InvalidArgumentException
      */
     public function testExpiresAfterInvalidThrowsException()
     {
