+++ test/Psr/CacheItemPool/MemoryIntegrationTest.php
@@ -1,0 +1,26 @@
+<?php
+/**
+ * @see       https://github.com/zendframework/zend-cache for the canonical source repository
+ * @copyright Copyright (c) 2018 Zend Technologies USA Inc. (http://www.zend.com)
+ * @license   https://github.com/zendframework/zend-cache/blob/master/LICENSE.md New BSD License
+ */
+
+namespace ZendTest\Cache\Psr\CacheItemPool;
+
+use PHPUnit\Framework\TestCase;
+use Zend\Cache\Psr\CacheItemPool\CacheItemPoolAdapter;
+use Zend\Cache\StorageFactory;
+
+class MemoryIntegrationTest extends TestCase
+{
+    /**
+     * The memory adapter calculates the TTL on reading which violates PSR-6
+     *
+     * @expectedException \Zend\Cache\Psr\CacheItemPool\CacheException
+     */
+    public function testAdapterNotSupported()
+    {
+        $storage = StorageFactory::adapterFactory('memory');
+        new CacheItemPoolAdapter($storage);
+    }
+}
