+++ test/Psr/CacheItemPool/MongoDbIntegrationTest.php
@@ -1,25 +1,20 @@
 <?php
 /**
- * Zend Framework (http://framework.zend.com/).
- *
- * @link      http://github.com/zendframework/zend-cache for the canonical source repository
- * @copyright Copyright (c) 2016 Zend Technologies USA Inc. (http://www.zend.com)
- * @license   http://framework.zend.com/license/new-bsd New BSD License
+ * @see       https://github.com/zendframework/zend-cache for the canonical source repository
+ * @copyright Copyright (c) 2018 Zend Technologies USA Inc. (http://www.zend.com)
+ * @license   https://github.com/zendframework/zend-cache/blob/master/LICENSE.md New BSD License
  */
 
-namespace ZendTest\Cache\Psr;
+namespace ZendTest\Cache\Psr\CacheItemPool;
 
 use Cache\IntegrationTests\CachePoolTest;
-use Zend\Cache\Psr\CacheItemPoolAdapter;
-use Zend\Cache\Storage\Adapter\Memcached;
+use Zend\Cache\Psr\CacheItemPool\CacheItemPoolAdapter;
+use Zend\Cache\Storage\Adapter\MongoDb;
 use Zend\Cache\StorageFactory;
 use Zend\Cache\Exception;
 use Zend\ServiceManager\Exception\ServiceNotCreatedException;
 
-/**
- * @require extension memcached
- */
-class MemcachedIntegrationTest extends CachePoolTest
+class MongoDbIntegrationTest extends CachePoolTest
 {
     /**
      * Backup default timezone
@@ -28,14 +23,14 @@ class MemcachedIntegrationTest extends CachePoolTest
     private $tz;
 
     /**
-     * @var Memcached
+     * @var MongoDb
      */
     private $storage;
 
     protected function setUp()
     {
-        if (! getenv('TESTS_ZEND_CACHE_MEMCACHED_ENABLED')) {
-            $this->markTestSkipped('Enable TESTS_ZEND_CACHE_MEMCACHED_ENABLED to run this test');
+        if (! getenv('TESTS_ZEND_CACHE_MONGODB_ENABLED')) {
+            $this->markTestSkipped('Enable TESTS_ZEND_CACHE_MONGODB_ENABLED to run this test');
         }
 
         // set non-UTC timezone
@@ -58,20 +53,12 @@ class MemcachedIntegrationTest extends CachePoolTest
 
     public function createCachePool()
     {
-        $host = getenv('TESTS_ZEND_CACHE_MEMCACHED_HOST');
-        $port = getenv('TESTS_ZEND_CACHE_MEMCACHED_PORT');
-
-        $options = [
-            'resource_id' => __CLASS__
-        ];
-        if ($host && $port) {
-            $options['servers'] = [[$host, $port]];
-        } elseif ($host) {
-            $options['servers'] = [[$host]];
-        }
-
         try {
-            $storage = StorageFactory::adapterFactory('memcached', $options);
+            $storage = StorageFactory::adapterFactory('mongodb', [
+                'server'     => getenv('TESTS_ZEND_CACHE_MONGODB_CONNECTSTRING'),
+                'database'   => getenv('TESTS_ZEND_CACHE_MONGODB_DATABASE'),
+                'collection' => getenv('TESTS_ZEND_CACHE_MONGODB_COLLECTION'),
+            ]);
 
             $deferredSkippedMessage = sprintf(
                 '%s storage doesn\'t support driver deferred',
