+++ test/Psr/CacheItemPool/WinCacheIntegrationTest.php
@@ -1,21 +1,23 @@
 <?php
 /**
- * Zend Framework (http://framework.zend.com/).
- *
- * @link      http://github.com/zendframework/zend-cache for the canonical source repository
- * @copyright Copyright (c) 2016 Zend Technologies USA Inc. (http://www.zend.com)
- * @license   http://framework.zend.com/license/new-bsd New BSD License
+ * @see       https://github.com/zendframework/zend-cache for the canonical source repository
+ * @copyright Copyright (c) 2018 Zend Technologies USA Inc. (http://www.zend.com)
+ * @license   https://github.com/zendframework/zend-cache/blob/master/LICENSE.md New BSD License
  */
 
-namespace ZendTest\Cache\Psr;
+namespace ZendTest\Cache\Psr\CacheItemPool;
 
 use Cache\IntegrationTests\CachePoolTest;
-use Zend\Cache\Psr\CacheItemPoolAdapter;
+use Zend\Cache\Psr\CacheItemPool\CacheItemPoolAdapter;
+use Zend\Cache\Storage\Adapter\WinCache;
 use Zend\Cache\StorageFactory;
 use Zend\Cache\Exception;
 use Zend\ServiceManager\Exception\ServiceNotCreatedException;
 
-class ZendServerDiskIntegrationTest extends CachePoolTest
+/**
+ * @requires extension wincache
+ */
+class WinCacheIntegrationTest extends CachePoolTest
 {
     /**
      * Backup default timezone
@@ -23,10 +25,15 @@ class ZendServerDiskIntegrationTest extends CachePoolTest
      */
     private $tz;
 
+    /**
+     * @var WinCache
+     */
+    private $storage;
+
     protected function setUp()
     {
-        if (! getenv('TESTS_ZEND_CACHE_ZEND_SERVER_ENABLED')) {
-            $this->markTestSkipped('Enable TESTS_ZEND_CACHE_ZEND_SERVER_ENABLED to run this test');
+        if (! getenv('TESTS_ZEND_CACHE_WINCACHE_ENABLED')) {
+            $this->markTestSkipped('Enable TESTS_ZEND_CACHE_WINCACHE_ENABLED to run this test');
         }
 
         // set non-UTC timezone
@@ -40,8 +47,8 @@ class ZendServerDiskIntegrationTest extends CachePoolTest
     {
         date_default_timezone_set($this->tz);
 
-        if (function_exists('zend_disk_cache_clear')) {
-            zend_disk_cache_clear();
+        if ($this->storage) {
+            $this->storage->flush();
         }
 
         parent::tearDown();
@@ -50,7 +57,7 @@ class ZendServerDiskIntegrationTest extends CachePoolTest
     public function createCachePool()
     {
         try {
-            $storage = StorageFactory::adapterFactory('zendserverdisk');
+            $storage = StorageFactory::adapterFactory('wincache');
 
             $deferredSkippedMessage = sprintf(
                 '%s storage doesn\'t support driver deferred',
