+++ test/Psr/SimpleCache/ApcIntegrationTest.php
@@ -9,7 +9,7 @@ namespace ZendTest\Cache\Psr\SimpleCache;
 
 use Cache\IntegrationTests\SimpleCacheTest;
 use Zend\Cache\Exception\ExtensionNotLoadedException;
-use Zend\Cache\Psr\SimpleCacheDecorator;
+use Zend\Cache\Psr\SimpleCache\SimpleCacheDecorator;
 use Zend\Cache\StorageFactory;
 use Zend\ServiceManager\Exception\ServiceNotCreatedException;
