+++ test/Psr/SimpleCache/ExtMongoDbIntegrationTest.php
@@ -9,7 +9,7 @@ namespace ZendTest\Cache\Psr\SimpleCache;
 
 use Cache\IntegrationTests\SimpleCacheTest;
 use MongoDB\Client;
-use Zend\Cache\Psr\SimpleCacheDecorator;
+use Zend\Cache\Psr\SimpleCache\SimpleCacheDecorator;
 use Zend\Cache\Storage\Adapter\ExtMongoDbOptions;
 use Zend\Cache\StorageFactory;
 use Zend\Cache\Exception;
