+++ test/Psr/SimpleCache/MongoDbIntegrationTest.php
@@ -8,7 +8,7 @@
 namespace ZendTest\Cache\Psr\SimpleCache;
 
 use Cache\IntegrationTests\SimpleCacheTest;
-use Zend\Cache\Psr\SimpleCacheDecorator;
+use Zend\Cache\Psr\SimpleCache\SimpleCacheDecorator;
 use Zend\Cache\Storage\Adapter\MongoDb;
 use Zend\Cache\StorageFactory;
 use Zend\Cache\Exception;
